requirements.txt.txt
# --- Web/API stack ---
fastapi==0.123.4
uvicorn==0.38.0
gunicorn==21.2.0
starlette==0.50.0

# --- Async/HTTP core -----
anyio==4.12.0
sniffio==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1

# --- Networking/requests ---
requests==2.32.5
urllib3==2.5.0
idna==3.11
certifi==2025.11.12
charset-normalizer==3.4.4

# --- Models / Typing / Config ---
pydantic==2.12.5
pydantic_core==2.41.5
typing_extensions==4.15.0
typing-inspection==0.4.2
packaging>=23,<26
python-dotenv==1.2.1

# --- Azure SDKs / Identity ---
azure-core==1.36.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
msal==1.34.0
msal-extensions==1.3.1
isodate==0.7.2

# --- Redis / Caching / Disk ---
redis==7.1.0
diskcache==5.6.3

# --- Databases ---
psycopg2==2.9.11
aioodbc==0.5.0
asyncpg==0.31.0
pyodbc==5.3.0

# --- Time / Dates ---
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3

# --- Numeric / Data ---
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0

# --- Autogen / OpenAI / Tokenizers ---
openai>=1.40.0
tiktoken==0.12.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext==0.7.5
pyautogen==0.7.5

# --- Telemetry ---
opentelemetry-api==1.38.0

# --- Utilities ---
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
asyncer==0.0.8
distro==1.9.0
docker==7.1.0
executing==2.2.1
fast-depends==2.4.12
importlib_metadata==8.7.0
jsonref==1.1.0
nest-asyncio==1.6.0
parso==0.8.5
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==5.29.5
psutil==7.1.3
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
regex==2025.11.3
six==1.17.0
stack-data==0.6.3
termcolor==3.2.0
tornado==6.5.2
tqdm==4.67.1
wcwidth==0.2.14
websockets==14.2
zipp==3.23.0

# --- Jupyter / Dev Tools ---
asttokens==3.0.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
greenlet==3.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
pyzmq==27.1.0

# --- SQLAlchemy ---
SQLAlchemy==2.0.45

# --- Misc ---
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
marklassian==0.1.0

# Ragas and compatible dependencies for Python 3.11
ragas==0.4.3

# Required LangChain compatibility to prevent version mismatches
langchain-core>=1.2,<1.3
langchain-openai>=1.1,<1.2

# Additional for RAG
azure-search-documents==11.6.0 
langchain-text-splitters==1.1.0
pdfplumber==0.11.8
python-multipart==0.0.22
